skip lost+found

+++ components/content-service/pkg/initializer/initializer.go
@@ -315,7 +315,11 @@ func InitializeWorkspace(ctx context.Context, location string, remoteStorage sto
 			return src, xerrors.Errorf("cannot clean workspace folder: %w", err)
 		}
 		for _, f := range fs {
-			path := filepath.Join(location, f.Name())
+			fname := f.Name()
+			if fname == "lost+found" {
+				continue
+			}
+			path := filepath.Join(location, fname)
 			err := os.RemoveAll(path)
 			if err != nil {
 				return src, xerrors.Errorf("cannot clean workspace folder: %w", err)
